Unit testing with the Goal controller and changed the Goal record slightly

+++ src/main/java/bino/consistent/grind/goals/Goal.java
@@ -1,7 +1,7 @@
 package bino.consistent.grind.goals;
 
 public record Goal(
-    long id,
+    Integer id,
     String goalName
 ) {
